feat: Add responsive landing page and move todo app
This commit introduces a new landing page for the application.

Key changes:
- Replaced the previous content of the root page (`/`) with a new, modern landing page.
- The landing page includes a hero section, feature highlights, and a call to action.
- It is styled using Tailwind CSS and is responsive across different screen sizes.
- Images and icons have been added for better visual appeal, including SVGs from the public directory and placeholder icons.
- The original todo list application has been moved to the `/todos` route, ensuring existing functionality remains accessible.
- Links to the `/todos` page are provided in the landing page's header, hero section, and CTA.

Testing:
- Jest and React Testing Library have been set up for the project.
- A test suite (`app/__tests__/page.test.tsx`) has been added for the new landing page.
- Tests cover the rendering of major sections, headings, buttons, links (verifying `href` attributes), and images.
- All 9 tests pass, confirming the landing page structure and basic functionality.
