+++ jest.setup.js
@@ -1,0 +1,2 @@
+// jest.setup.js
+import '@testing-library/jest-dom'; // Provides .toBeInTheDocument() and other matchers
